# docs/intent-hub-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Unified Intent Hub",
    "version": "0.2.0",
    "description": "Generated route inventory for master profile. Fixed compatibility namespaces are stable; root aliases follow API_COMPAT_PROFILE."
  },
  "paths": {
...
    "/diagnostics/overlap": {
      "get": {
        "operationId": "analyze_all_overlaps_get",
        "tags": [
          "master"
        ],
        "responses": {
          "200": {
            "description": "Completed response; see API.md for endpoint-specific statuses and errors."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/diagnostics/umap": {
      "get": {
        "operationId": "diagnostics_umap_get",
        "tags": [
          "master"
        ],
        "responses": {
          "200": {
            "description": "Completed response; see API.md for endpoint-specific statuses and errors."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/diagnostics/repair": {
      "post": {
        "operationId": "get_repair_suggestions_post",
        "tags": [
          "master"
        ],
        "responses": {
          "200": {
            "description": "Completed response; see API.md for endpoint-specific statuses and errors."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/diagnostics/apply-repair": {
      "post": {
        "operationId": "apply_repair_post",
        "tags": [
          "master"
        ],
        "responses": {
          "200": {
            "description": "Completed response; see API.md for endpoint-specific statuses and errors."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/settings": {
      "get": {
        "operationId": "get_settings_get",
        "tags": [
          "master"
        ],
        "responses": {
          "200": {
            "description": "Completed response; see API.md for endpoint-specific statuses and errors."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "operationId": "update_settings_post",
        "tags": [
          "master"
        ],
        "responses": {
          "200": {
            "description": "Completed response; see API.md for endpoint-specific statuses and errors."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/agents": {
      "get": {
        "operationId": "bupt_root_bupt_agents_get",
        "tags": [
          "bupt"
        ],
        "responses": {
          "200": {
            "description": "Completed response; see API.md for endpoint-specific statuses and errors."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "operationId": "bupt_root_bupt_create_agent_post",
        "tags": [
          "bupt"
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreate"
              }
            }
          }
        }
      }
    },
...
    "/agents/pull": {
      "post": {
        "operationId": "bupt_root_bupt_pull_agents_post",
        "tags": [
          "bupt"
        ],
        "responses": {
          "200": {
            "description": "Completed response; see API.md for endpoint-specific statuses and errors."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/vectors/sync": {
      "post": {
        "operationId": "bupt_root_bupt_sync_vectors_post",
        "tags": [
          "bupt"
        ],
        "responses": {
          "200": {
            "description": "Completed response; see API.md for endpoint-specific statuses and errors."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/sync": {
      "post": {
        "operationId": "bupt_root_bupt_sync_post",
        "tags": [
          "bupt"
        ],
        "responses": {
          "200": {
            "description": "Completed response; see API.md for endpoint-specific statuses and errors."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/sync/status": {
      "get": {
        "operationId": "bupt_root_bupt_sync_status_get",
        "tags": [
          "bupt"
        ],
        "responses": {
          "200": {
            "description": "Completed response; see API.md for endpoint-specific statuses and errors."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/route": {
      "post": {
        "operationId": "bupt_root_bupt_route_post",
        "tags": [
          "bupt"
        ],
        "responses": {
          "200": {
            "description": "BUPT success/data/error envelope; data.agents contains agent details and nullable score."
          },
          "503": {
            "description": "Routing components warming or unavailable; retry shortly."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          }
        }
      }
    },
    "/collections": {
      "get": {
        "operationId": "bupt_root_bupt_collections_get",
        "tags": [
          "bupt"
        ],
        "responses": {
          "200": {
            "description": "Completed response; see API.md for endpoint-specific statuses and errors."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "operationId": "bupt_root_bupt_create_collection_post",
        "tags": [
          "bupt"
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/collections/restore": {
      "post": {
        "operationId": "bupt_root_bupt_restore_collection_post",
        "tags": [
          "bupt"
        ],
        "responses": {
          "200": {
            "description": "Completed response; see API.md for endpoint-specific statuses and errors."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/diagnostics/merge": {
      "post": {
        "operationId": "bupt_root_bupt_diagnostics_merge_post",
        "tags": [
          "bupt"
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeAgentsRequest"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "RouteRequest": {
        "properties": {
          "query": {
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "RouteRequest",
        "type": "object"
      },
...
      "AgentCreate": {
        "properties": {
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          },
          "utterances": {
            "items": {
              "type": "string"
            },
            "title": "Utterances",
            "type": "array"
          },
          "negative_samples": {
            "items": {
              "type": "string"
            },
            "title": "Negative Samples",
            "type": "array"
          },
          "score_threshold": {
            "default": 0.8,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score Threshold",
            "type": "number"
          },
          "negative_threshold": {
            "default": 0.95,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Negative Threshold",
            "type": "number"
          }
        },
        "required": [
          "title"
        ],
        "title": "AgentCreate",
        "type": "object"
      },
...
      "MergeAgentsRequest": {
        "properties": {
          "source_agent_id": {
            "title": "Source Agent Id",
            "type": "integer"
          },
          "target_agent_id": {
            "title": "Target Agent Id",
            "type": "integer"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "source_agent_id",
          "target_agent_id",
          "title"
        ],
        "title": "MergeAgentsRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}